requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
coqui-tts>=0.27.2
soundfile>=0.12.1
numpy>=1.23